Moved favicon/title tests to their own file
